+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+## [0.5.3] - 2023-05-22
+
 ### Fixed
 
 - Failure to read module parameters without types, such as `loop`
@@ -186,7 +188,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release. No notable features
 
 <!-- next-url -->
-[Unreleased]: https://github.com/DianaNites/linapi/compare/v0.5.2...HEAD
+[Unreleased]: https://github.com/DianaNites/linapi/compare/v0.5.3...HEAD
+[0.5.3]: https://github.com/DianaNites/linapi/compare/v0.5.2...v0.5.3
 [0.5.2]: https://github.com/DianaNites/linapi/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/DianaNites/linapi/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/DianaNites/linapi/compare/v0.4.4...v0.5.0